Rich markdown text editor:
Implemented a watered down but still sophisticated form of Krajee.

A quick fix for an error of an article authour inability to edit their
post was also fixed

I also made sure that in this commit, All files are loaded,
that is, no error was produced on the browser console.

Lastly, the advert image distortion was fixed